download/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
playwright==1.49.1
playwright-stealth==1.0.6
requests==2.32.3
pydantic==2.10.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
urllib3==1.26.20
setuptools==75.6.0
weasyprint
cairocffi
tinycss2